added initial commit-support & skipping
Instead of creating a `git diff` of two commits, niftypee now uses `git
show HEAD` to list all files affected by the last commit. This allows
to synchronize the first commit.

Niftypee now skips unnecessary (and time consuming) steps for the
generation of the directory-batches if the preceding put- or
delete-batch is empty.